Create separate multichain module
