+++ src/sync/multichain/models/body/stream-body.ts
