+++ src/sync/multichain/models/body/transaction-body.ts
